made navbar
added home link, booking button and fontawesome dropdown placeholder
