+++ src/components/header/header.module.scss
@@ -1,0 +1,13 @@
+.HomeLink {
+  color: white;
+  text-decoration: none;
+}
+
+.Header {
+  margin: 0;
+  max-width: 100%;
+  padding: 1rem 1rem;
+  border: 1px solid red;
+  display: flex;
+  justify-content: space-between;
+}
